+++ Scripts/Components/Other/DSC_ADD_CursorLock.cs
@@ -1,0 +1,14 @@
+﻿using Unity.Entities;
+using UnityEngine;
+
+namespace DSC.Actor.DOTS
+{
+    [GenerateAuthoringComponent]  
+    public struct DSC_ADD_CursorLock : IComponentData
+    {
+        public CursorLockMode m_eLockMode;
+
+        [HideInInspector]
+        public CursorLockMode m_ePreviousMode;
+    }
+}