+++ Scripts/Systems/Empty/DSC_ADS_GroundCheck.cs
@@ -1,0 +1,13 @@
+﻿using Unity.Entities;
+
+namespace DSC.Actor.DOTS
+{
+    [UpdateInGroup(typeof(InitializationSystemGroup))]
+    public sealed class DSC_ADS_GroundCheck : SystemBase
+    {
+        protected override void OnUpdate()
+        {
+
+        }
+    }
+}