+++ Scripts/Systems/Empty/DSC_ADS_Move_Pre.cs
@@ -1,0 +1,12 @@
+﻿using Unity.Entities;
+
+namespace DSC.Actor.DOTS
+{
+    public sealed class DSC_ADS_Move_Pre : SystemBase
+    {
+        protected override void OnUpdate()
+        {
+            
+        }
+    }
+}