+++ Scripts/Systems/Hybrid/Debug/DSC_ADS_Debug_Rigidbody_Velocity.cs
@@ -1,0 +1,21 @@
+﻿using Unity.Entities;
+using UnityEngine;
+
+namespace DSC.Actor.DOTS
+{
+    public sealed class DSC_ADS_Debug_Rigidbody_Velocity : SystemBase
+    {
+        protected override void OnUpdate()
+        {
+            Entities.WithAll<DSC_ADT_Debug_Rigidbody_Velocity>()
+                .ForEach((Rigidbody hRigid) =>
+            {
+                if (hRigid)
+                {
+                    Debug.Log("Velocity : " + hRigid.velocity,hRigid.gameObject);
+                }
+            }).WithoutBurst()
+            .Run();
+        }
+    }
+}